+++ test/functional/utils_test.js
@@ -1,12 +1,12 @@
-var request = require('request');
 var sinon = require('sinon');
 var Promise = require('bluebird');
 var _ = require('underscore');
 var packageJson = require('../../package.json');
 var fs = require('fs');
 const { smartSheetURIs } = require('../..');
+var axios = require('axios');
 
-var requestor = require('../../lib/utils/httpRequestor').create({request: request});
+var requestor = require('../../lib/utils/httpRequestor').create({request: axios});
 
 var sample = {
   name : 'name'
@@ -198,18 +198,20 @@ describe('Utils Unit Tests', function() {
     describe('#Successful request', function() {
       var requestStub = null;
       var stubbedRequestor = require('../../lib/utils/httpRequestor')
-        .create({request: request, handleResponse: () => ({content: true})});
+        .create({request: axios, handleResponse: () => ({ content: true })});
 
       beforeEach(() => {
-        requestStub = sinon.stub(request, 'getAsync');
+        requestStub = sinon.stub(axios, 'get');
         var mockResponse = {
-          statusCode: 200,
+          status: 200,
           headers: {
             'content-type':'application/json;charset=UTF-8'
+          },
+          data: {
+            hello: "world"
           }
         };
-        var mockBody = '{"hello":"world"}';
-        requestStub.returns(Promise.resolve([mockResponse, mockBody]));
+        requestStub.returns(Promise.resolve(mockResponse));
       });
 
       afterEach(() => {
@@ -231,11 +233,11 @@ describe('Utils Unit Tests', function() {
     describe('#Error on request', function() {
       var requestStub = null;
       var stubbedRequestor = require('../../lib/utils/httpRequestor')
-        .create({request: request, handleResponse: () => ({content: true})});
+        .create({request: axios, handleResponse: () => ({ content: true })});
       var mockBody;
 
       beforeEach(() => {
-        requestStub = sinon.stub(request, 'getAsync');
+        requestStub = sinon.stub(axios, 'get');
         var mockResponse = {
           statusCode: 403,
           headers: {
@@ -269,7 +271,7 @@ describe('Utils Unit Tests', function() {
       var spyGet;
 
       beforeEach(() => {
-        spyGet = sinon.spy(request, 'getAsync');
+        spyGet = sinon.spy(axios, 'get');
       });
 
       afterEach(() => {
@@ -284,20 +286,20 @@ describe('Utils Unit Tests', function() {
           'User-Agent': `smartsheet-javascript-sdk/${EXPECTED_VERSION}`
         };
         requestor.get(sampleRequest);
-        spyGet.args[0][0].headers.Authorization.should.equal(sampleHeaders.Authorization);
-        spyGet.args[0][0].headers.Accept.should.equal(sampleHeaders.Accept);
-        spyGet.args[0][0].headers['Content-Type'].should.equal(sampleHeaders['Content-Type']);
-        spyGet.args[0][0].headers['User-Agent'].should.equal(sampleHeaders['User-Agent']);
+        spyGet.args[0][1].headers.Authorization.should.equal(sampleHeaders.Authorization);
+        spyGet.args[0][1].headers.Accept.should.equal(sampleHeaders.Accept);
+        spyGet.args[0][1].headers['Content-Type'].should.equal(sampleHeaders['Content-Type']);
+        spyGet.args[0][1].headers['User-Agent'].should.equal(sampleHeaders['User-Agent']);
       });
 
       it('url sent to request should match given', () => {
         requestor.get(sampleRequest);
-        spyGet.args[0][0].url.should.equal('https://api.smartsheet.com/2.0/URL');
+        spyGet.args[0][0].should.equal('https://api.smartsheet.com/2.0/URL');
       });
 
       it('queryString sent to request should match given', () => {
         requestor.get(sampleRequestWithQueryParameters);
-        spyGet.args[0][0].qs.should.equal(sampleRequestWithQueryParameters.queryParameters);
+        spyGet.args[0][1].params.should.equal(sampleRequestWithQueryParameters.queryParameters);
       });
     });
 
@@ -305,7 +307,7 @@ describe('Utils Unit Tests', function() {
       var requestStub = null;
       var handleResponseStub = sinon.stub();
       var stubbedRequestor = require('../../lib/utils/httpRequestor')
-        .create({request: request, handleResponse: handleResponseStub});
+        .create({request: axios, handleResponse: handleResponseStub});
       var sampleRequestForRetry = null;
 
       function givenGetReturnsError() {
@@ -330,7 +332,7 @@ describe('Utils Unit Tests', function() {
       }
 
       beforeEach(() => {
-        requestStub = sinon.stub(request, 'getAsync');
+        requestStub = sinon.stub(axios, 'get');
         sampleRequestForRetry = _.extend({}, sampleRequest);
         sampleRequestForRetry.maxRetryDurationMillis = 30;
         sampleRequestForRetry.calcRetryBackoff = function (numRetry) {return Math.pow(3, numRetry);};
@@ -377,10 +379,10 @@ describe('Utils Unit Tests', function() {
       var requestStub = null;
 
       var stubbedRequestor = require('../../lib/utils/httpRequestor')
-        .create({request: request, handleResponse: () => ({content: true})});
+        .create({request: axios, handleResponse: () => ({content: true})});
 
       beforeEach(() => {
-        requestStub = sinon.stub(request, 'postAsync');
+        requestStub = sinon.stub(axios, 'post');
         var mockResponse = {
           statusCode: 200,
           headers: {
@@ -413,10 +415,10 @@ describe('Utils Unit Tests', function() {
       var mockBody = {error:true};
 
       var stubbedRequestor = require('../../lib/utils/httpRequestor')
-        .create({request: request, handleResponse: () => ({content: true})});
+        .create({request: axios, handleResponse: () => ({content: true})});
 
       beforeEach(() => {
-        requestStub = sinon.stub(request, 'postAsync');
+        requestStub = sinon.stub(axios, 'post');
         requestStub.returns(Promise.reject(mockBody));
       });
 
@@ -443,7 +445,7 @@ describe('Utils Unit Tests', function() {
       var spyPost;
 
       beforeEach(() => {
-        spyPost = sinon.spy(request, 'postAsync');
+        spyPost = sinon.spy(axios, 'post');
       });
 
       afterEach(() => {
@@ -458,25 +460,25 @@ describe('Utils Unit Tests', function() {
           'User-Agent': `smartsheet-javascript-sdk/${EXPECTED_VERSION}`
         };
         requestor.post(sampleRequest);
-        spyPost.args[0][0].headers.Authorization.should.equal(sampleHeaders.Authorization);
-        spyPost.args[0][0].headers.Accept.should.equal(sampleHeaders.Accept);
-        spyPost.args[0][0].headers['Content-Type'].should.equal(sampleHeaders['Content-Type']);
-        spyPost.args[0][0].headers['User-Agent'].should.equal(sampleHeaders['User-Agent']);
+        spyPost.args[0][2].headers.Authorization.should.equal(sampleHeaders.Authorization);
+        spyPost.args[0][2].headers.Accept.should.equal(sampleHeaders.Accept);
+        spyPost.args[0][2].headers['Content-Type'].should.equal(sampleHeaders['Content-Type']);
+        spyPost.args[0][2].headers['User-Agent'].should.equal(sampleHeaders['User-Agent']);
       });
 
       it('url sent to request should match given', () => {
         requestor.post(sampleRequest);
-        spyPost.args[0][0].url.should.equal('https://api.smartsheet.com/2.0/URL');
+        spyPost.args[0][0].should.equal('https://api.smartsheet.com/2.0/URL');
       });
 
       it('queryString sent to request should match given', () => {
         requestor.post(sampleRequestWithQueryParameters);
-        spyPost.args[0][0].qs.should.equal(sampleRequestWithQueryParameters.queryParameters);
+        spyPost.args[0][2].params.should.equal(sampleRequestWithQueryParameters.queryParameters);
       });
 
       it('body sent to request should match given', () => {
         requestor.post(sampleRequestWithQueryParameters);
-        spyPost.args[0][0].body.should.equal(JSON.stringify(sampleRequestWithQueryParameters.body));
+        spyPost.args[0][1].should.equal(sampleRequestWithQueryParameters.body);
       });
     });
 
@@ -485,7 +487,7 @@ describe('Utils Unit Tests', function() {
       var handleResponseStub = sinon.stub();
 
       var stubbedRequestor = require('../../lib/utils/httpRequestor')
-        .create({request: request, handleResponse: handleResponseStub});
+        .create({request: axios, handleResponse: handleResponseStub});
 
       var sampleRequestForRetry;
 
@@ -511,7 +513,7 @@ describe('Utils Unit Tests', function() {
       }
 
       beforeEach(() => {
-        requestStub = sinon.stub(request, 'postAsync');
+        requestStub = sinon.stub(axios, 'post');
 
         sampleRequestForRetry = _.extend({}, sampleRequest);
         sampleRequestForRetry.maxRetryDurationMillis = 30;
@@ -559,10 +561,10 @@ describe('Utils Unit Tests', function() {
       var requestStub = null;
 
       var stubbedRequestor = require('../../lib/utils/httpRequestor')
-        .create({request: request, handleResponse: () => ({content: true})});
+        .create({request: axios, handleResponse: () => ({content: true})});
 
       beforeEach(() => {
-        requestStub = sinon.stub(request, 'putAsync');
+        requestStub = sinon.stub(axios, 'put');
         var mockResponse = {
           statusCode: 200,
           headers: {
@@ -597,10 +599,10 @@ describe('Utils Unit Tests', function() {
       var mockBody = {error: true};
 
       var stubbedRequestor = require('../../lib/utils/httpRequestor')
-        .create({request: request, handleResponse: () => ({content: true})});
+        .create({request: axios, handleResponse: () => ({content: true})});
 
       beforeEach(() => {
-        stub = sinon.stub(request, 'putAsync');
+        stub = sinon.stub(axios, 'put');
         var mockResponse = {
           statusCode: 403,
           headers: {
@@ -633,7 +635,7 @@ describe('Utils Unit Tests', function() {
       var spyPut;
 
       beforeEach(() => {
-        spyPut = sinon.spy(request, 'putAsync');
+        spyPut = sinon.spy(axios, 'put');
       });
 
       afterEach(() => {
@@ -648,25 +650,25 @@ describe('Utils Unit Tests', function() {
           'User-Agent': `smartsheet-javascript-sdk/${EXPECTED_VERSION}`
         };
         requestor.put(sampleRequest);
-        spyPut.args[0][0].headers.Authorization.should.equal(sampleHeaders.Authorization);
-        spyPut.args[0][0].headers.Accept.should.equal(sampleHeaders.Accept);
-        spyPut.args[0][0].headers['Content-Type'].should.equal(sampleHeaders['Content-Type']);
-        spyPut.args[0][0].headers['User-Agent'].should.equal(sampleHeaders['User-Agent']);
+        spyPut.args[0][2].headers.Authorization.should.equal(sampleHeaders.Authorization);
+        spyPut.args[0][2].headers.Accept.should.equal(sampleHeaders.Accept);
+        spyPut.args[0][2].headers['Content-Type'].should.equal(sampleHeaders['Content-Type']);
+        spyPut.args[0][2].headers['User-Agent'].should.equal(sampleHeaders['User-Agent']);
       });
 
       it('url sent to request should match given', () => {
         requestor.put(sampleRequest);
-        spyPut.args[0][0].url.should.equal('https://api.smartsheet.com/2.0/URL');
+        spyPut.args[0][0].should.equal('https://api.smartsheet.com/2.0/URL');
       });
 
       it('queryString sent to request should match given', () => {
         requestor.put(sampleRequestWithQueryParameters);
-        spyPut.args[0][0].qs.should.equal(sampleRequestWithQueryParameters.queryParameters);
+        spyPut.args[0][2].params.should.equal(sampleRequestWithQueryParameters.queryParameters);
       });
 
       it('body sent to request should match given', () => {
         requestor.put(sampleRequestWithQueryParameters);
-        spyPut.args[0][0].body.should.equal(JSON.stringify(sampleRequestWithQueryParameters.body));
+        spyPut.args[0][1].should.equal(sampleRequestWithQueryParameters.body);
       });
     });
 
@@ -675,7 +677,7 @@ describe('Utils Unit Tests', function() {
       var handleResponseStub = sinon.stub();
 
       var stubbedRequestor = require('../../lib/utils/httpRequestor')
-        .create({request: request, handleResponse: handleResponseStub});
+        .create({request: axios, handleResponse: handleResponseStub});
 
       var sampleRequestForRetry = null;
 
@@ -701,7 +703,7 @@ describe('Utils Unit Tests', function() {
       }
 
       beforeEach(() => {
-        requestStub = sinon.stub(request, 'putAsync');
+        requestStub = sinon.stub(axios, 'put');
 
         sampleRequestForRetry = _.extend({}, sampleRequest);
         sampleRequestForRetry.maxRetryDurationMillis = 30;
@@ -749,10 +751,10 @@ describe('Utils Unit Tests', function() {
       var requestStub = null;
 
       var stubbedRequestor = require('../../lib/utils/httpRequestor')
-        .create({request: request, handleResponse: () => ({content: true})});
+        .create({request: axios, handleResponse: () => ({content: true})});
 
       beforeEach(() => {
-        requestStub = sinon.stub(request, 'delAsync');
+        requestStub = sinon.stub(axios, 'delete');
         var mockResponse = {
           statusCode: 200,
           headers: {
@@ -788,10 +790,10 @@ describe('Utils Unit Tests', function() {
       var mockBody = {error: true};
 
       var stubbedRequestor = require('../../lib/utils/httpRequestor')
-        .create({request: request, handleResponse: () => ({content: true})});
+        .create({request: axios, handleResponse: () => ({content: true})});
 
       beforeEach(() => {
-        requestStub = sinon.stub(request, 'delAsync');
+        requestStub = sinon.stub(axios, 'delete');
         var mockResponse = {
           statusCode: 403,
           headers: {
@@ -824,7 +826,7 @@ describe('Utils Unit Tests', function() {
       var spyPut;
 
       beforeEach(() => {
-        spyPut = sinon.spy(request, 'delAsync');
+        spyPut = sinon.spy(axios, 'delete');
       });
 
       afterEach(() => {
@@ -839,20 +841,20 @@ describe('Utils Unit Tests', function() {
           'User-Agent': `smartsheet-javascript-sdk/${EXPECTED_VERSION}`
         };
         requestor.delete(sampleRequest);
-        spyPut.args[0][0].headers.Authorization.should.equal(sampleHeaders.Authorization);
-        spyPut.args[0][0].headers.Accept.should.equal(sampleHeaders.Accept);
-        spyPut.args[0][0].headers['Content-Type'].should.equal(sampleHeaders['Content-Type']);
-        spyPut.args[0][0].headers['User-Agent'].should.equal(sampleHeaders['User-Agent']);
+        spyPut.args[0][1].headers.Authorization.should.equal(sampleHeaders.Authorization);
+        spyPut.args[0][1].headers.Accept.should.equal(sampleHeaders.Accept);
+        spyPut.args[0][1].headers['Content-Type'].should.equal(sampleHeaders['Content-Type']);
+        spyPut.args[0][1].headers['User-Agent'].should.equal(sampleHeaders['User-Agent']);
       });
 
       it('url sent to request should match given', () => {
         requestor.delete(sampleRequest);
-        spyPut.args[0][0].url.should.equal('https://api.smartsheet.com/2.0/URL');
+        spyPut.args[0][0].should.equal('https://api.smartsheet.com/2.0/URL');
       });
 
       it('queryString sent to request should match given', () => {
         requestor.delete(sampleRequestWithQueryParameters);
-        spyPut.args[0][0].qs.should.equal(sampleRequestWithQueryParameters.queryParameters);
+        spyPut.args[0][1].params.should.equal(sampleRequestWithQueryParameters.queryParameters);
       });
     });
 
@@ -861,7 +863,7 @@ describe('Utils Unit Tests', function() {
       var handleResponseStub = sinon.stub();
 
       var stubbedRequestor = require('../../lib/utils/httpRequestor')
-        .create({request: request, handleResponse: handleResponseStub});
+        .create({request: axios, handleResponse: handleResponseStub});
 
       var sampleRequestForRetry;
 
@@ -887,7 +889,7 @@ describe('Utils Unit Tests', function() {
       }
 
       beforeEach(() => {
-        requestStub = sinon.stub(request, 'delAsync');
+        requestStub = sinon.stub(axios, 'delete');
 
         sampleRequestForRetry = _.extend({}, sampleRequest);
         sampleRequestForRetry.maxRetryDurationMillis = 30;